add SQLRunningContext#run()
