+++ springboot-starter-data-authorization/pom.xml
@@ -6,7 +6,7 @@
     <parent>
         <groupId>com.codingapi.springboot</groupId>
         <artifactId>springboot-parent</artifactId>
-        <version>3.3.32</version>
+        <version>3.3.33</version>
     </parent>
 
     <artifactId>springboot-starter-data-authorization</artifactId>
