+++ springboot-starter-data-authorization/src/main/java/com/codingapi/springboot/authorization/jdbc/SQLRunningState.java
@@ -1,0 +1,14 @@
+package com.codingapi.springboot.authorization.jdbc;
+
+public enum SQLRunningState {
+
+    /**
+     * 拦截
+     */
+    INTERCEPT,
+
+    /**
+     * 不拦截
+     */
+    UN_INTERCEPT;
+}
