+++ springboot-starter-data-authorization/src/main/java/com/codingapi/springboot/authorization/jdbc/proxy/ConnectionProxy.java
@@ -1,7 +1,6 @@
 package com.codingapi.springboot.authorization.jdbc.proxy;
 
 import com.codingapi.springboot.authorization.interceptor.SQLInterceptState;
-import com.codingapi.springboot.authorization.interceptor.SQLRunningContext;
 
 import java.sql.*;
 import java.util.Map;
