+++ springboot-starter-data-fast/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <artifactId>springboot-parent</artifactId>
         <groupId>com.codingapi.springboot</groupId>
-        <version>3.3.32</version>
+        <version>3.3.33</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
