+++ springboot-starter-flow/pom.xml
@@ -6,7 +6,7 @@
     <parent>
         <artifactId>springboot-parent</artifactId>
         <groupId>com.codingapi.springboot</groupId>
-        <version>3.3.32</version>
+        <version>3.3.33</version>
     </parent>
 
     <name>springboot-starter-flow</name>
